Unify stack around NO_NETWORK_TESTING / RELEASE TESTING
This makes several important changes:

1. The maximum amount of coverage is automatically provided if
anyone specifies RELEASE_TESTING or AUTOMATED_TESTING in their
environment.  AUTOMATED_TESTING is employed on CPAN Testers, but *not*
on typical end users machines.

2. If the environement variable NO_NETWORK_TESTING is set, network
tests will not be run, regardless of the above setting. This is to
standardize on a widely used variable defined for this purpose, so
that the people who are fine with running automated tests, but not
network tests, can do that as they see fit.

3. Tests that require threads will self-disable when threads aren't
available. This is mostly to avoid the need to mess around with setting
variables just to run thread-enabled tests.

The variable ONLINE_TESTING is still used internally, but not used in
notices, mostly for compatibility with any existing infrastructure that
the author may have been using that set this, and for compatibility
with any downstreams who have set this flag in response to past
messages.

Closes: https://github.com/csirtgadgets/LWPx-ParanoidAgent/issues/27